refactor : 클래스 명, PaymentConfig 수정
바뀐 내용 :
1)PaymentClient -> PaymentApiSerializer
API 직렬화, 역직렬화만 담당하는 클래스라는 인터페이스명

2)PaymentConfig -> PaymentProperites로 변경
config보다는 Properties 성격
